+++ docs/web_modules/LayoutContainer/index.js
@@ -43,6 +43,15 @@ export default class Layout extends Component {
             { src: "https://cdn.polyfill.io/v2/polyfill.min.js" },
           ] }
         />
+
+        { /* meta viewport safari/chrome/edge */ }
+        <Helmet
+          meta={ [ {
+            name: "viewport", content: "width=device-width, initial-scale=1",
+          } ] }
+        />
+        <style>{ "@-ms-viewport { width: device-width; }" }</style>
+
         <div className={ styles.layout }>
           <Header />
           <div className={ styles.content }>